requirements.txt
gradio
langchain
langchain_community
langchain_core
pandas
pyprojroot
python-dotenv
PyYAML
SQLAlchemy
openai
fastapi
uvicorn
google-generativeai>=0.7.0

